requirements.txt
cryptography>=41.0.0
customtkinter>=5.2.0
Pillow>=10.0.0 